+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- collections & playlists screens
+
 ## [v0.6.3] - 2026-08-23
 
 ### Added